refactor: PersonsCard переименован в PersonCard
