+++ src/components/ui/person-card/index.ts
@@ -1,0 +1,1 @@
+export * from './person-card';
