--- src/components/ui/persons-card/index.ts
@@ -1,1 +1,0 @@
-export * from './persons-card';
